md-to-pdf-converter/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "md-to-pdf-converter"
version = "1.0.0"
description = "A professional markdown to PDF converter with customizable themes"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["markdown", "pdf", "converter", "document", "report", "theme"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Utilities",
]
dependencies = [
    "markdown2>=2.4.0",
    "weasyprint>=52.5",
    "Pillow>=8.0.0",
    "click>=8.0.0",
    "colorama>=0.4.4",
]

[project.urls]
Homepage = "https://github.com/yourusername/md-to-pdf-converter"
Repository = "https://github.com/yourusername/md-to-pdf-converter"
BugTracker = "https://github.com/yourusername/md-to-pdf-converter/issues"

[project.scripts]
md2pdf = "cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["src"] by default)
include = ["core*", "cli*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests*"]  # exclude packages matching these glob patterns (empty by default)

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.css"]